backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.7
yfinance==0.2.22
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
gunicorn==21.2.0

